Task: Add project structure.
