+++ .travis.yml
@@ -1,8 +1,7 @@
 language: node_js
 node_js:
+  - "10"
   - "8"
-  - "6"
-  - "4"
 addons:
   code_climate:
     repo_token: 2d851f8f3a9348ac4f43262305037f80a730c2660fda50af8ae4d445fd89333b
