+++ lib/bind.js
